Added ortho_cluster3/OG_stats/
Some formatting changes were made to corresponding scripts in ortho_cluster1/ and ortho_cluster2/, but more importantly, the pipeline was modified in two ways. In the first change, OGs2hits/ and OGs2hsps/ only record the gene pairs rather than each individual hit. The files are then joined on gene-gene pairs rather than index. This creates major storage savings. In the second change, hits and HSPs were properly filtered with the reciprocity information. Before all hits associated with a gene-gene pair were added if the gene-gene pair was part of an OG. This is not correct since hits are only considered part of the gene graph if they are reciprocal in the first place.