Increased effective sequence number and fixed paths
Previously the sequence number was the number of sequences in the alignment. In one rare case this
was not enough to shift the default emission priors towards the observed alignment, causing a large
region to be spuriously and poorly aligned. Increasing the effective sequence number, essentially the
weight on the observed alignment as opposed to the priors, fixed this issue.