Removed duplicated code when plotting norm1 and norm2 MSAs by consolidating in for loop
Previously when the MSAs were plotted, the duplicated code was short enough that it felt cleaner to copy and paste rather than use a for loop.
However, the duplicated portion in trim_plot.py was large enough to warrant a proper for loop, and I went ahead and fixed it in the other places too.