Modified trim.py to extend margin search region in both directions
In the first version, some bounds were incorrectly chosen because the columns with the greatest change were in the core region.
Thus, the margins are extended towards the interior to capture these cases. The search is stopped when the gradient decreases below some
threshold. This is set higher than the one used for the exteriors since the confidence of a state tends to level off rapidly in the interior
whereas the exteriors can be more ambiguous and gradually approach zero. Finally, the bounds are chosen as the columns that maximize the
product of the delta and the gradient. Usually the HMM identifies the transition region within a few columns with high accuracy, but I was
skeptical that the column with the greatest change in the raw number of gaps would ALWAYS be the one with the largest gradient. By taking the
product of the two, any idiosyncracies in the gradient are more than compensated by the large delta corresponding to the correct column.
Conversely when the raw deltas are misleading due to a short insertion in one sequence in a nearby column, the gradient compensates by setting
the product near zero since, as mentioned before, the HMM is typically flexible enough to ignore these cases when calculating the posterior.