Adjusted trimming algorithm to reflect better decodings
The updated decodings more rapidly toggle between states and tend to be more confident in their predictions.
In light of this, I adjusted the method in which the exterior margin is adjusted. The first insight was that even though
I want to be fairly selective when selecting the core regions, they are often surrounded by fairly high confidence shoulders,
which should be included due to their proximity to the high-confidence core. Thus, as long as the shoulders are above some lower
threshold (POST_LOW), they will be included in the margin. The second condition (the gradient being below a certain threshold)
is unchanged. However, by including it second in the OR expression, it will only be checked if the posterior has past POST_LOW.
This ensures that convex regions above POST_LOW aren't falsely excluded.