Updated insertion_hmm training scripts to use paths to old alignments
I tried to update the training data to use the new alignments. However, the majority of the first 5 or 6 alignments were not identical,
meaning I had to update large portions of the labels. This was not trivial in some cases since the new alignments sometimes created
entirely new regions of labels (rather than just shifting the columns around slightly). Since I was very satisfied with the original HMM
parameters, I decided to not update the labels. Training data is training data, so it shouldn't matter that the HMMs are trained on
alignments that are a 'dead-end' for further analysis. I will have to update the deletion_hmm training data, however, since those labels
are based on alignments after the insertions have been trimmed. This training data is much smaller, though, so it will be much easier to
make those changes.