Added note about meta alignments not necessarily producing consistent outputs
Although the random seed is set consistently, the inputs are not sorted and may
differ in order depending on the conditions. (One of the issues is dictionary
iteration not being guaranteed. The other issue is os.listdir not guaranteeing
consistent order either.) This is left as is since fixing it would change the
order regardless.