Moved calculation of rates outside of likelihood calculations
This is only refactoring code, so it shouldn't change the results. Logically the rate calculations are closer to the
construction of the partitions, so it makes more sense for this to happen when the partition data is extracted from
the IQ-TREE output. The only major change is the rates also incorporate the 'speed' parameter which further scales
the rate matrix (or the length of the branches). Before the rate was multiplied by the speed immediately before
calling get_conditional. With this change, the likelihood calculations are identical between scripts.